+++ mtg_proxy_printer/ui/main_window.py
@@ -18,8 +18,9 @@ import typing
 
 
 from PyQt5.QtCore import pyqtSlot, pyqtSignal, QStringListModel
-from PyQt5.QtGui import QCloseEvent, QKeySequence
+from PyQt5.QtGui import QCloseEvent, QKeySequence, QDesktopServices
 from PyQt5.QtWidgets import QApplication, QMessageBox, QProgressBar, QAction, QWidget, QToolBar, QLabel
+
 
 from mtg_proxy_printer.missing_images_manager import MissingImagesManager
 from mtg_proxy_printer.card_info_downloader import CardInfoDownloader
@@ -373,12 +374,15 @@ class MainWindow(*inherits_from_ui_file_with_name(f"main_window")):
             )
 
     def show_application_update_available_message_box(self, newer_version: str):
-        QMessageBox.information(
-            self, "Update available",
-            f"An application update is available: Version {newer_version}\n\n"
-            f"You are currently using version {mtg_proxy_printer.meta_data.__version__}.",
-            QMessageBox.Ok, QMessageBox.Ok
-        )
+        if QMessageBox.question(
+                self, "Application update available. Visit website?",
+                f"An application update is available: Version {newer_version}\n"
+                f"You are currently using version {mtg_proxy_printer.meta_data.__version__}.\n\n"
+                f"Open the {mtg_proxy_printer.meta_data.PROGRAMNAME} website in your webbrowser "
+                f"to download the new version?",
+                QMessageBox.Yes | QMessageBox.No, QMessageBox.No
+                ) == QMessageBox.Yes:
+            QDesktopServices.openUrl(mtg_proxy_printer.meta_data.DOWNLOAD_WEB_PAGE)
 
     def show_card_data_update_available_message_box(self, estimated_card_count: int):
         if QMessageBox.question(